# api/openapi/v1.yaml
openapi: "3.1.0"
info:
  title: Custos API
  version: "0.1.0"
  description: |
    Multi-tenant Slurm gateway / HPC control plane. Conventions per
    docs/api.md: JSON, snake_case fields, RFC 3339 UTC timestamps,
    keyset pagination via opaque cursors, UUIDv7 identifiers.
servers:
  - url: /api/v1
paths:
...
  /tenants/{tenant}:
    get:
      summary: Get a tenant
      operationId: getTenant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      responses:
        "200":
          description: The tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          description: Unauthenticated.
          content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
        "404":
          description: Tenant not found (or not visible to the caller).
          content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    patch:
      summary: Update a tenant
      operationId: updateTenant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated tenant.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          description: Unauthenticated.
          content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
        "404":
          description: Tenant not found.
          content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
        "409":
          description: Version or state conflict.
          content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
    delete:
      summary: Request tenant deletion
      description: |
        Marks the tenant `deleting` and enqueues asynchronous deletion
        (platform.manage). The tenant row is retained as a tombstone;
        slugs are never reused.
      operationId: deleteTenant
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantSlug"
      responses:
        "202":
          description: Deletion accepted.
        "401":
          description: Unauthenticated.
          content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
        "403":
          description: Not a platform admin.
          content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
        "404":
          description: Tenant not found.
          content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
        "409":
          description: Already deleting or deleted.
          content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}
...
  /tenants/{tenant}/secret-connectors:
    parameters:
      - $ref: "#/components/parameters/TenantSlug"
    get:
      operationId: listSecretConnectors
      summary: List tenant secret connectors
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Connector list, content: {application/json: {schema: {type: object}}}}
    post:
      operationId: createSecretConnector
      summary: Create a tenant secret connector
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateSecretConnector"}}}}
      responses:
        "201": {description: Connector created, content: {application/json: {schema: {$ref: "#/components/schemas/SecretConnector"}}}}
        "422": {description: Invalid or unavailable provider, content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
...
  /tenants/{tenant}/secret-references:
    parameters:
      - $ref: "#/components/parameters/TenantSlug"
    get:
      operationId: listSecretReferences
      security: [{bearerAuth: []}]
      responses: {"200": {description: Reference list, content: {application/json: {schema: {type: object}}}}}
    post:
      operationId: createSecretReference
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: "#/components/schemas/CreateSecretReference"}}}}
      responses: {"201": {description: Reference created, content: {application/json: {schema: {$ref: "#/components/schemas/SecretReference"}}}}, "422": {description: Invalid reference}}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
            request_id:
              type: string
            details:
              type: array
              items:
                type: object
                required: [field, reason]
                properties:
                  field:
                    type: string
                  reason:
                    type: string
...
    Tenant:
      type: object
      required: [id, slug, name, state, settings, version, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        state:
          type: string
          enum: [provisioning, active, suspended, deleting, deleted]
        settings:
          type: object
        openbao_namespace:
          type: string
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TenantUpdate:
      type: object
      required: [version]
      properties:
        name:
          type: string
        settings:
          type: object
        state:
          type: string
          enum: [active, suspended]
        version:
          type: integer
...
    SecretConnector:
      type: object
      required: [id, tenant_id, name, kind, state, config, version]
      properties:
        id: {type: string, format: uuid}
        tenant_id: {type: string, format: uuid}
        name: {type: string}
        kind: {type: string, enum: [platform-openbao, openbao]}
        state: {type: string, enum: [active, disabled]}
        config: {type: object, additionalProperties: true}
        credential_ref: {$ref: "#/components/schemas/SecretRef"}
        version: {type: integer, format: int64}
    CreateSecretConnector:
      type: object
      required: [name, kind, config, credential]
      properties:
        name: {type: string}
        kind: {type: string, enum: [openbao]}
        config: {type: object, additionalProperties: true}
        credential:
          type: object
          additionalProperties: {type: string}
          writeOnly: true
    SecretReference:
      type: object
      required: [id, tenant_id, name, connector_id, namespace, mount, path, key, kind, allowed_uses, version]
      properties:
        id: {type: string, format: uuid}
        tenant_id: {type: string, format: uuid}
        owner_id: {type: [string, "null"], format: uuid}
        project_id: {type: [string, "null"], format: uuid}
        name: {type: string}
        connector_id: {type: string, format: uuid}
        namespace: {type: string}
        mount: {type: string}
        path: {type: string}
        key: {type: string}
        secret_version: {type: [integer, "null"], minimum: 1}
        kind: {type: string, enum: [ssh_key, slurm_token, api_token, generic, storage_credential]}
        allowed_uses: {type: array, items: {type: string}}
        version: {type: integer, format: int64}
    CreateSecretReference:
      type: object
      required: [name, path, key, kind]
      properties:
        name: {type: string}
        connector: {type: string, default: default}
        owner_id: {type: string, format: uuid}
        project_id: {type: string, format: uuid}
        mount: {type: string}
        path: {type: string}
        key: {type: string}
        secret_version: {type: integer, minimum: 1}
        kind: {type: string, enum: [ssh_key, slurm_token, api_token, generic, storage_credential]}
        allowed_uses: {type: array, items: {type: string}}
    SecretRef:
      type: object
      description: Reference to a secret value (docs/secrets.md); never the value.
      required: [provider, path]
      properties:
        provider:
          type: string
          enum: [file, openbao]
        namespace:
          type: string
          description: OpenBao namespace; required for openbao references.
        mount:
          type: string
          description: OpenBao KV v2 mount; required for openbao references.
        path:
          type: string
        key:
          type: string
          description: Field within the secret document.
        version:
          type: integer
          minimum: 1
          description: Optional pinned OpenBao KV version.
...
  parameters:
...
    TenantSlug:
      name: tenant
      in: path
      required: true
      description: Tenant slug or uuid.
      schema:
        type: string
...